amplify/functions/agent-discovery-scheduler/requirements.txt
boto3>=1.26.0
python-dateutil>=2.8.2
